#194 払戻テキストのパース処理を RaceResultPayoutParser に切り出し
StoreAction が抱えていた payout テキストのパース機能（TICKET_TYPE_MAP /
REQUIRED_TYPES 定数、parse / parseHorseNumbers / parseAmount /
parsePopularity / validateAllTypesPresent メソッド）を専用サービス
クラスに分離。StoreAction は DI で受け取って委譲する形に変更し、
保存処理と入力解釈の責務を疎結合にした。

外部インターフェースは不変。空テキスト時の早期 throw のみ承認済みの
テストに合わせて補強している。
